Added a few more tests, changed db.
- Added Database ctor to specify name/path
- Added a check to `removeGameBy*` functions so as to not remove non-existant fields
- Added seperate db for test cases to use

+++ Source/Database.cpp
@@ -9,6 +9,13 @@ Database::Database()
     db.setDatabaseName("ascension.db");
 }
 
+Database::Database(QString path)
+    : db(QSqlDatabase::addDatabase("QSQLITE"))
+{
+    db.setHostName("localhost");
+    db.setDatabaseName(path);
+}
+
 bool Database::init()
 {
     bool status = db.open();
@@ -19,21 +26,19 @@ bool Database::init()
     }
 
     QSqlQuery createQuery(db);
-    createQuery.exec("CREATE TABLE IF NOT EXISTS games(ID INTEGER PRIMARY KEY ASC, GAMENAME TEXT NOT NULL, GAMEDIRECTORY TEXT NOT NULL, GAMEEXECUTABLE TEXT NOT NULL, ARGUMENTS TEXT NOT NULL);");
-
-    return true;
+    return createQuery.exec("CREATE TABLE IF NOT EXISTS games(ID INTEGER PRIMARY KEY ASC, GAMENAME TEXT NOT NULL, GAMEDIRECTORY TEXT NOT NULL, GAMEEXECUTABLE TEXT NOT NULL, ARGUMENTS TEXT NOT NULL);");
 }
 
 bool Database::reset()
 {
     QSqlQuery query(db);
-    return query.exec("DROP TABLES *");
+    return query.exec("DROP TABLE IF EXISTS games");
 }
 
 bool Database::addGame(QString gameName, QString gameDirectory, QString executablePath, QString arguments)
 {
     QSqlQuery query(db);
-    query.prepare("INSERT INTO GAMES(GAMENAME, GAMEDIRECTORY, GAMEEXECUTABLE, ARGUMENTS) VALUES (:gameName, :gameDirectory, :executablePath, :arguments);");
+    query.prepare("INSERT OR IGNORE INTO GAMES(GAMENAME, GAMEDIRECTORY, GAMEEXECUTABLE, ARGUMENTS) VALUES (:gameName, :gameDirectory, :executablePath, :arguments);");
     query.bindValue(":gameName", gameName);
     query.bindValue(":gameDirectory", gameDirectory);
     query.bindValue(":executablePath", executablePath);
@@ -43,18 +48,32 @@ bool Database::addGame(QString gameName, QString gameDirectory, QString executab
 
 bool Database::removeGameById(unsigned int id)
 {
-    QSqlQuery query(db);
-    query.prepare("DELETE FROM GAMES WHERE ID = :id;");
-    query.bindValue(":id", id);
-    return query.exec();
+    if (std::get<0>(isExistant(id)))
+    {
+        QSqlQuery query(db);
+        query.prepare("DELETE FROM games WHERE ID = :id;");
+        query.bindValue(":id", id);
+        return query.exec();
+    }
+    else
+    {
+        return false;
+    }
 }
 
 bool Database::removeGameByName(QString name)
 {
-    QSqlQuery query(db);
-    query.prepare("DELETE FROM GAMES WHERE GAMENAME = :name;");
-    query.bindValue(":name", name);
-    return query.exec();
+    if (std::get<0>(isExistant(name)))
+    {
+        QSqlQuery query(db);
+        query.prepare("DELETE FROM GAMES WHERE GAMENAME = :name;");
+        query.bindValue(":name", name);
+        return query.exec();
+    }
+    else
+    {
+        return false;
+    }
 }
 
 Game Database::getGameById(unsigned int id) { return std::get<1>(isExistant(id)); }
